db_flask/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
alembic==1.13.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Pillow==10.2.0
WTForms~=3.2.1
Faker~=37.3.0
tqdm~=4.67.1